Move all datastore shims into separate “adapter” modules that consume the public API only
