+++ datastore/core/binarystore.py
@@ -26,6 +26,10 @@ class Datastore:
 	databases where fields should be decomposed into columns), particularly to
 	support queries efficiently.
 	"""
+
+	# Some possibly useful types (assigned at the end of this file)
+	ADAPTER_T: type
+	RECEIVE_T: type
 	
 	# Main API. Datastore implementations MUST implement these methods.
 	
@@ -39,7 +43,7 @@ class Datastore:
 		You **must** exhaust or manually close the returned iterable to ensure
 		that possibly associated resources, like open file descriptors, are
 		free'd.
-		
+
 		Arguments
 		---------
 		key
@@ -143,8 +147,8 @@ class Datastore:
 			return True
 		except KeyError:
 			return False
-	
-	
+
+
 
 class NullDatastore(Datastore):
 	"""Stores nothing, but conforms to the API. Useful to test with."""
@@ -169,7 +173,7 @@ class NullDatastore(Datastore):
 
 class DictDatastore(Datastore):
 	"""Simple straw-man in-memory datastore backed by nested dicts."""
-	
+
 	_items: typing.Dict[str, typing.Dict[key_.Key, bytes]]
 
 	def __init__(self):
@@ -331,8 +335,8 @@ class Adapter(Datastore):
 		  key: Key naming the object to remove.
 		"""
 		await self.child_datastore.delete(key)
-
-
+	
+	
 	async def contains(self, key: key_.Key) -> bool:
 		"""Returns whether any data named by `key` exists
 		
@@ -363,469 +367,8 @@ class Adapter(Datastore):
 		return await self.child_datastore.query(query)
 
 
-
-class LoggingAdapter(Adapter):
-	"""Wraps a datastore with a logging shim."""
-	
-	def __init__(self, child_datastore, logger=None):
-		if not logger:
-			import logging
-			logger = logging
-		
-		self.logger = logger
-		
-		super().__init__(child_datastore)
-	
-	
-	async def get(self, key: key_.Key) -> util.stream.ReceiveStream:
-		"""Return the object named by key or None if it does not exist.
-		   LoggingDatastore logs the access.
-		"""
-		self.logger.info('%s: get %s' % (self, key))
-		value = await super().get(key)
-		self.logger.debug('%s: %s' % (self, value))
-		return value
-	
-	
-	async def _put(self, key: key_.Key, value: util.stream.ReceiveStream) -> None:
-		"""Stores the object `value` named by `key`self.
-		   LoggingDatastore logs the access.
-		"""
-		self.logger.info('%s: put %s' % (self, key))
-		self.logger.debug('%s: %s' % (self, value))
-		await super()._put(key, value)
-	
-	
-	async def delete(self, key: key_.Key) -> None:
-		"""Removes the object named by `key`.
-		   LoggingDatastore logs the access.
-		"""
-		self.logger.info('%s: delete %s' % (self, key))
-		await super().delete(key)
-	
-	
-	async def contains(self, key: key_.Key) -> bool:
-		"""Returns whether the object named by `key` exists.
-		   LoggingDatastore logs the access.
-		"""
-		self.logger.info('%s: contains %s' % (self, key))
-		return await super().contains(key)
-	
-	
-	async def query(self, query: query_.Query) -> query_.Cursor:
-		"""Returns an iterable of objects matching criteria expressed in `query`.
-		   LoggingDatastore logs the access.
-		"""
-		self.logger.info('%s: query %s' % (self, query))
-		return await super().query(query)
-
-
-
-class KeyTransformAdapter(Adapter):
-	"""Represents a simple DatastoreAdapter that applies a transform on all incoming
-	   keys. For example:
-
-		>>> import datastore.core
-		>>> def transform(key):
-		...   return key.reverse
-		...
-		>>> ds = datastore.DictDatastore()
-		>>> kt = datastore.KeyTransformDatastore(ds, keytransform=transform)
-		None
-		>>> ds.put(datastore.Key('/a/b/c'), 'abc')
-		>>> ds.get(datastore.Key('/a/b/c'))
-		'abc'
-		>>> kt.get(datastore.Key('/a/b/c'))
-		None
-		>>> kt.get(datastore.Key('/c/b/a'))
-		'abc'
-		>>> ds.get(datastore.Key('/c/b/a'))
-		None
-	"""
-	
-	def __init__(self, *args, keytransform=(lambda k: k), **kwargs):
-		"""Initializes KeyTransformDatastore with `keytransform` function."""
-		self.keytransform = keytransform
-		super().__init__(*args, **kwargs)
-	
-	
-	async def get(self, key: key_.Key) -> util.stream.ReceiveStream:
-		"""Return the object named by keytransform(key)."""
-		return await super().get(self.keytransform(key))
-	
-	
-	async def _put(self, key: key_.Key, value: util.stream.ReceiveStream) -> None:
-		"""Stores the object names by keytransform(key)."""
-		await super().put(self.keytransform(key), value)
-	
-	
-	async def delete(self, key: key_.Key) -> None:
-		"""Removes the object named by keytransform(key)."""
-		await super().delete(self.keytransform(key))
-	
-	
-	async def contains(self, key: key_.Key) -> bool:
-		"""Returns whether the object named by key is in this datastore."""
-		return await super().contains(self.keytransform(key))
-	
-	
-	async def query(self, query: query_.Query) -> query_.Cursor:
-		"""Returns a sequence of objects matching criteria expressed in `query`"""
-		query = query.copy()
-		query.key = self.keytransform(query.key)
-		return await super().query(query)
-
-
-class LowercaseKeyAdapter(KeyTransformAdapter):
-	"""Represents a simple DatastoreAdapter that lowercases all incoming keys.
-	   For example:
-
-		>>> import datastore.core
-		>>> ds = datastore.DictDatastore()
-		>>> ds.put(datastore.Key('hello'), 'world')
-		>>> ds.put(datastore.Key('HELLO'), 'WORLD')
-		>>> ds.get(datastore.Key('hello'))
-		'world'
-		>>> ds.get(datastore.Key('HELLO'))
-		'WORLD'
-		>>> ds.get(datastore.Key('HeLlO'))
-		None
-		>>> lds = datastore.LowercaseKeyDatastore(ds)
-		>>> lds.get(datastore.Key('HeLlO'))
-		'world'
-		>>> lds.get(datastore.Key('HeLlO'))
-		'world'
-		>>> lds.get(datastore.Key('HeLlO'))
-		'world'
-	"""
-
-	def __init__(self, *args, **kwargs):
-		"""Initializes KeyTransformDatastore with keytransform function."""
-		super().__init__(*args, keytransform=self.lowercase_key, **kwargs)
-	
-	
-	@classmethod
-	def lowercase_key(cls, key: key_.Key) -> key_.Key:
-		"""Returns a lowercased `key`."""
-		return key_.Key(str(key).lower())
-
-
-
-class NamespaceAdapter(KeyTransformAdapter):
-	"""Represents a simple DatastoreAdapter that namespaces all incoming keys.
-	   For example:
-
-		>>> import datastore.core
-		>>>
-		>>> ds = datastore.DictDatastore()
-		>>> ds.put(datastore.Key('/a/b'), 'ab')
-		>>> ds.put(datastore.Key('/c/d'), 'cd')
-		>>> ds.put(datastore.Key('/a/b/c/d'), 'abcd')
-		>>>
-		>>> nd = datastore.NamespaceDatastore('/a/b', ds)
-		>>> nd.get(datastore.Key('/a/b'))
-		None
-		>>> nd.get(datastore.Key('/c/d'))
-		'abcd'
-		>>> nd.get(datastore.Key('/a/b/c/d'))
-		None
-		>>> nd.put(datastore.Key('/c/d'), 'cd')
-		>>> ds.get(datastore.Key('/a/b/c/d'))
-		'cd'
-	"""
-
-	def __init__(self, namespace, *args, **kwargs):
-		"""Initializes NamespaceDatastore with `key` namespace."""
-		self.namespace = key_.Key(namespace)
-		super().__init__(*args, keytransform = self.namespace_key, **kwargs)
-	
-	
-	def namespace_key(self, key: key_.Key) -> key_.Key:
-		"""Returns a namespaced `key`: namespace.child(key)."""
-		return self.namespace.child(key)
-
-
-
-class NestedPathAdapter(KeyTransformAdapter):
-	"""Represents a simple DatastoreAdapter that shards/namespaces incoming keys.
-
-	Incoming keys are sharded into nested namespaces. The idea is to use the key
-	name to separate into nested namespaces. This is akin to the directory
-	structure that ``git`` uses for objects. For example:
-
-		>>> import datastore.core
-		>>>
-		>>> ds = datastore.DictDatastore()
-		>>> np = datastore.NestedPathDatastore(ds, depth=3, length=2)
-		>>>
-		>>> np.put(datastore.Key('/abcdefghijk'), 1)
-		>>> np.get(datastore.Key('/abcdefghijk'))
-		1
-		>>> ds.get(datastore.Key('/abcdefghijk'))
-		None
-		>>> ds.get(datastore.Key('/ab/cd/ef/abcdefghijk'))
-		1
-		>>> np.put(datastore.Key('abc'), 2)
-		>>> np.get(datastore.Key('abc'))
-		2
-		>>> ds.get(datastore.Key('/ab/ca/bc/abc'))
-		2
-	"""
-
-	_default_depth:  int = 3
-	_default_length: int = 2
-	_default_keyfn:  typing.Callable[[key_.Key], str] = staticmethod(lambda key: key.name)
-
-	def __init__(self, *args,
-	             depth: int = None,
-	             length: int = None,
-	             keyfn: typing.Callable[[key_.Key], str] = None,
-	             **kwargs):
-		"""Initializes KeyTransformDatastore with keytransform function.
-
-		Arguments
-		---------
-		depth
-			The nesting level depth (e.g. 3 => /1/2/3/123); default: 3
-		length:
-			The nesting level length (e.g. 2 => /12/123456); default: 2
-		keyfn:
-			A function that maps key paths to the name they should be stored as
-		"""
-
-		# assign the nesting variables
-		self.nest_depth  = depth  if depth  is not None else self._default_depth
-		self.nest_length = length if length is not None else self._default_length
-		self.nest_keyfn  = keyfn  if keyfn  is not None else self._default_keyfn
-
-		super().__init__(*args, keytransform=self.nest_key, **kwargs)
-	
-	
-	async def query(self, query: query_.Query) -> query_.Cursor:
-		# Requires supporting * operator on queries.
-		raise NotImplementedError()
-	
-	
-	def nest_key(self, key: key_.Key) -> key_.Key:
-		"""Returns a nested `key`."""
-		
-		nest = self.nest_keyfn(key)
-		
-		# if depth * length > len(key.name), we need to pad.
-		mult = 1 + int(self.nest_depth * self.nest_length / len(nest))
-		nest = nest * mult
-		
-		pref = key_.Key(self.nested_path(nest, self.nest_depth, self.nest_length))
-		return pref.child(key)
-	
-	
-	@staticmethod
-	def nested_path(path: str, depth: int, length: int) -> str:
-		"""Returns a nested version of `basename`, using the starting characters.
-		
-		For example:
-		
-			>>> NestedPathDatastore.nested_path('abcdefghijk', 3, 2)
-			'ab/cd/ef'
-			>>> NestedPathDatastore.nested_path('abcdefghijk', 4, 2)
-			'ab/cd/ef/gh'
-			>>> NestedPathDatastore.nested_path('abcdefghijk', 3, 4)
-			'abcd/efgh/ijk'
-			>>> NestedPathDatastore.nested_path('abcdefghijk', 1, 4)
-			'abcd'
-			>>> NestedPathDatastore.nested_path('abcdefghijk', 3, 10)
-			'abcdefghij/k'
-		"""
-		components = [path[n:n + length] for n in range(0, len(path), length)]
-		components = components[:depth]
-		return '/'.join(components)
-
-
-
-class DatastoreCollectionMixin:
-	"""Represents a collection of datastores."""
-
-	def __init__(self, stores: typing.Collection[Datastore] = []):
-		"""Initialize the datastore with any provided datastores."""
-		if not isinstance(stores, list):
-			stores = list(stores)
-		self._stores = stores
-
-	def get_datastore_at(self, index: int) -> Datastore:
-		"""Returns the datastore at `index`."""
-		return self._stores[index]
-
-	def append_datastore(self, store: Datastore) -> None:
-		"""Appends datastore `store` to this collection."""
-		self._stores.append(store)
-
-	def remove_datastore(self, store: Datastore) -> None:
-		"""Removes datastore `store` from this collection."""
-		self._stores.remove(store)
-
-	def insert_datastore(self, index: int, store: Datastore) -> None:
-		"""Inserts datastore `store` into this collection at `index`."""
-		self._stores.insert(index, store)
-
-
-class TieredAdapter(Adapter, DatastoreCollectionMixin):
-	"""Represents a hierarchical collection of datastores.
-
-	Each datastore is queried in order. This is helpful to organize access
-	order in terms of speed (i.e. read caches first).
-
-	Datastores should be arranged in order of completeness, with the most complete
-	datastore last, as it will handle query calls.
-
-	Semantics:
-	
-		* get      : returns first found value
-		* put      : writes through to all
-		* delete   : deletes through to all
-		* contains : returns first found value
-		* query    : queries bottom (most complete) datastore
-	"""
-
-	async def get(self, key: key_.Key) -> util.stream.ReceiveStream:
-		"""Return the object named by key. Checks each datastore in order."""
-		value: typing.Optional[util.stream.ReceiveStream] = None
-		exceptions: typing.List[KeyError] = []
-		for store in self._stores:
-			try:
-				value_: util.stream.ReceiveStream = await store.get(key)
-			except KeyError as exc:
-				exceptions.append(exc)
-			else:
-				value = value_
-				break
-		if value is None:
-			raise trio.MultiError(exceptions)
-		
-		# Add model to lower stores only
-		async with trio.open_nursery() as nursery:
-			for store2 in self._stores:
-				if store == store2:
-					break
-				nursery.call_soon(store2.put(key, value))
-		
-		return value
-	
-	
-	async def _put(self, key: key_.Key, value: util.stream.ReceiveStream) -> None:
-		"""Stores the object in all underlying datastores."""
-		async with trio.open_nursery() as nursery:
-			for store in self._stores:
-				nursery.call_soon(store.put(key, value))
-	
-	
-	async def delete(self, key: key_.Key) -> None:
-		"""Removes the object from all underlying datastores."""
-		async def swallow_key_error(coroutine):
-			try:
-				await coroutine
-			except KeyError:
-				pass
-		
-		async with trio.open_nursery() as nursery:
-			for store in self._stores:
-				nursery.call_soon(swallow_key_error(store.delete()))
-	
-	
-	async def query(self, query: query_.Query) -> query_.Cursor:
-		"""Returns a sequence of objects matching criteria expressed in `query`.
-		The last datastore will handle all query calls, as it has a (if not
-		the only) complete record of all objects.
-		"""
-		# queries hit the last (most complete) datastore
-		return self._stores[-1].query(query)
-	
-	
-	async def contains(self, key: key_.Key) -> bool:
-		"""Returns whether the object is in this datastore."""
-		for store in self._stores:
-			if await store.contains(key):
-				return True
-		return False
-
-
-
-class ShardedAdapter(Adapter, DatastoreCollectionMixin):
-	"""Represents a collection of datastore shards.
-	
-	A datastore is selected based on a sharding function.
-	Sharding functions should take a Key and return an integer.
-	
-	Caution
-	-------
-	Adding or removing datastores while mid-use may severely affect consistency.
-	Also ensure the order is correct upon initialization. While this is not as
-	important for caches, it is crucial for persistent datastores.
-	"""
-	
-	_shardingfn: typing.Callable[[key_.Key], int]
-	
-	
-	def __init__(self, stores: typing.Collection = [],
-	             shardingfn: typing.Callable[[key_.Key], int] = hash):
-		"""Initialize the datastore with any provided datastore."""
-		DatastoreCollectionMixin.__init__(self, stores)
-		self._shardingfn = shardingfn
-	
-	
-	def shard(self, key: key_.Key) -> int:
-		"""Returns the shard index to handle `key`, according to sharding fn."""
-		return self._shardingfn(key) % len(self._stores)
-	
-	
-	def get_sharded_datastore(self, key: key_.Key) -> Datastore:
-		"""Returns the shard to handle `key`."""
-		return self.get_datastore_at(self.shard(key))
-	
-	
-	async def get(self, key: key_.Key) -> util.stream.ReceiveStream:
-		"""Return the object named by key from the corresponding datastore."""
-		return await self.get_sharded_datastore(key).get(key)
-	
-	
-	async def _put(self, key: key_.Key, value: util.stream.ReceiveStream) -> None:
-		"""Stores the object to the corresponding datastore."""
-		await self.get_sharded_datastore(key).put(key, value)
-	
-	
-	async def delete(self, key: key_.Key) -> None:
-		"""Removes the object from the corresponding datastore."""
-		await self.get_sharded_datastore(key).delete(key)
-	
-	
-	async def contains(self, key: key_.Key) -> bool:
-		"""Returns whether the object is in this datastore."""
-		return await self.get_sharded_datastore(key).contains(key)
-	
-	
-	async def query(self, query: query_.Query) -> query_.Cursor:
-		"""Returns a sequence of objects matching criteria expressed in `query`"""
-		cursor = query_.Cursor(query, self._shard_query_generator(query))
-		cursor.apply_order()  # ordering sharded queries is expensive (no generator)
-		return cursor
-	
-	
-	def _shard_query_generator(self, query):
-		"""A generator that queries each shard in sequence."""
-		shard_query = query.copy()
-		
-		for shard in self._stores:
-			# yield all items matching within this shard
-			cursor = shard.query(shard_query)
-			for item in cursor:
-				yield item
-			
-			# update query with results of first query
-			shard_query.offset = max(shard_query.offset - cursor.skipped, 0)
-			if shard_query.limit:
-				shard_query.limit = max(shard_query.limit - cursor.returned, 0)
-				
-				if shard_query.limit <= 0:
-					break  # we're already done!
+Datastore.ADAPTER_T = Adapter
+Datastore.RECEIVE_T = util.stream.ReceiveStream
 
 
 """
